+++ app/models/carto/user_service.rb
@@ -73,12 +73,20 @@ module Carto
                                       .count
     end
 
+    def private_privacy_visualization_count
+      return 0 unless @user.id
+
+      Carto::VisualizationQueryBuilder.user_private_privacy_visualizations(@user)
+                                      .build
+                                      .count
+    end
+
     def all_visualization_count
       return 0 unless @user.id
 
       Carto::VisualizationQueryBuilder.user_all_visualizations(@user)
-        .build
-        .count
+                                      .build
+                                      .count
     end
 
     def twitter_imports_count(options={})